+++ composer.json
@@ -26,7 +26,7 @@
         "phpunit/phpunit": "4.7.*"
     },
     "autoload": {
-        "psr-0": {
+        "psr-4": {
             "SimpleSoftwareIO\\SMS\\": "src/"
         }
     },
